Optimizations phase 1
* Significant restructing to code base for better maintainability
* Groundwork of DPS update unit test which requires multiple requests
* Clean up of some imports
* Leverage new _constants.py
