+++ azext_iot/__init__.py
@@ -5,22 +5,26 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core import AzCommandsLoader
-from azext_iot._factory import iot_hub_service_factory
 from azure.cli.core.commands import CliCommandType
+from azext_iot._factory import iot_hub_service_factory, iot_service_provisioning_factory
 import azext_iot._help  # pylint: disable=unused-import
 
 
-iotext_custom = CliCommandType(
-    operations_tmpl='azext_iot.custom#{}',
+iothub_ops = CliCommandType(
+    operations_tmpl='azext_iot.operations.hub#{}',
     client_factory=iot_hub_service_factory
+)
+
+iotdps_ops = CliCommandType(
+    operations_tmpl='azext_iot.operations.dps#{}',
+    client_factory=iot_service_provisioning_factory
 )
 
 
 class IoTExtCommandsLoader(AzCommandsLoader):
     def __init__(self, cli_ctx=None):
         super(IoTExtCommandsLoader, self).__init__(cli_ctx=cli_ctx,
-                                                   min_profile='2017-03-10-profile',
-                                                   custom_command_type=iotext_custom)
+                                                   min_profile='2017-03-10-profile')
 
     def load_command_table(self, args):
         from azext_iot.commands import load_command_table
