+++ azext_iot/_constants.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------  
+
+import os
+
+EXTENSION_ROOT = os.path.dirname(os.path.abspath(__file__))
